image-processing-webapp-requirements.txt
Flask==2.0.1
Pillow==8.3.1
numpy==1.21.0
gunicorn==20.1.0
